refactorizado html del registro
