+++ src/main/java/vitalsanity/controller/PacienteController.java
@@ -31,14 +31,8 @@ public class PacienteController{
         return "paciente/ver-profesionales-medicos-autorizados";
     }
 
-    @GetMapping("/paciente-datos-personales-form")
-    public String datosPersonalesForm(Model model) {
-        return "paciente/completar-datos-personales";
-    }
-
     @GetMapping("/paciente-datos-residencia-form")
     public String datosResidenciaForm(Model model) {
         return "paciente/completar-datos-residencia";
     }
-
 }
