+++ src/main/java/vitalsanity/dto/RegistroData.java
@@ -1,0 +1,22 @@
+package vitalsanity.dto;
+
+public class RegistroData {
+    private String email;
+    private String contrasenya;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getContrasenya() {
+        return contrasenya;
+    }
+
+    public void setContrasenya(String contrasenya) {
+        this.contrasenya = contrasenya;
+    }
+}
