+++ src/main/resources/templates/guest-user/registro-form.html
@@ -14,7 +14,31 @@
             <h1 class="form-title">Formulario de registro</h1>
             <form>
                 <div class="form-group">
-                    <label class="form-label" for="email">Correo*</label>
+                    <label class="form-label" for="complete-name">Nombre completo*</label>
+                    <input type="email" id="complete-name" class="form-input" required placeholder="Pablo Hernández Ripoll">
+                </div>
+                <div class="form-group">
+                    <label class="form-label" for="birthdate">Fecha de nacimiento*</label>
+                    <input type="date" id="birthdate" class="form-input" required placeholder="06/09/11">
+                </div>
+                <div class="form-group">
+                    <label class="form-label" for="genre">Género*</label>
+                    <input type="text" id="genre" class="form-input" required placeholder="Hombre">
+                </div>
+                <div class="form-group">
+                    <label class="form-label" for="nifnie">NIF/NIE*</label>
+                    <input type="text" id="nifnie" class="form-input" required placeholder="43337475R">
+                </div>
+                <div class="form-group">
+                    <label class="form-label" for="phone">Móvil*</label>
+                    <input type="text" id="phone" class="form-input" required placeholder="+34 647239815">
+                </div>
+                <div class="form-group">
+                    <label class="form-label" for="country">País*</label>
+                    <input type="email" id="country" class="form-input" required placeholder="España">
+                </div>
+                <div class="form-group">
+                    <label class="form-label" for="email">Email*</label>
                     <input type="email" id="email" class="form-input" required placeholder="user@gmail.com">
                 </div>
                 <div class="form-group">
