--- src/main/resources/templates/paciente/completar-datos-personales.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Completar datos personales</title>
-    <div th:include="navbar :: head (titulo='Completar datos personales')"></div>
-    <link rel="stylesheet" th:href="@{/css/custom/paciente/completar-datos-personales.css}">
-</head>
-<body>
-<div th:replace="navbar::menuBar"></div>
-<div class="page-container">
-    <div class="login-container">
-        <h1 class="form-title">Complete sus datos personales</h1>
-        <form>
-            <div class="form-group">
-                <label class="form-label" for="complete-name">Nombre completo*</label>
-                <input type="email" id="complete-name" class="form-input" required placeholder="Pablo Hernández Ripoll">
-            </div>
-            <div class="form-group">
-                <label class="form-label" for="birthdate">Fecha de nacimiento*</label>
-                <input type="date" id="birthdate" class="form-input" required placeholder="06/09/11">
-            </div>
-            <div class="form-group">
-                <label class="form-label" for="genre">Género*</label>
-                <input type="text" id="genre" class="form-input" required placeholder="Hombre">
-            </div>
-            <div class="form-group">
-                <label class="form-label" for="nifnie">NIF/NIE*</label>
-                <input type="text" id="nifnie" class="form-input" required placeholder="43337475R">
-            </div>
-            <div class="form-group">
-                <label class="form-label" for="phone">Móvil*</label>
-                <input type="text" id="phone" class="form-input" required placeholder="+34 647239815">
-            </div>
-            <button type="submit" class="submit-button">Aceptar</button>
-        </form>
-    </div>
-</div>
-<div th:replace="navbar::javascript"></div>
-</body>
-</html>
